Extract LabelChooser
